fixed auth and validation

+++ app/src/main/java/com/zeuroux/launchly/gplay/GPlayService.kt
@@ -43,7 +43,7 @@ class GPlayService(
         try {
             val files = PurchaseHelper(authData()).purchase(
                 MINECRAFT_PACKAGE,
-                versionCode.toInt(),
+                versionCode,
                 0
             )
             files.mapNotNull { file ->
